Fix size on File object after optimizing
Since the size attribute/property on a File object is a cached_property, we need to make sure we update it after changing the data since it might already have cached the old size.

This becomes important because Google Cloud Storage's client library refuses to upload a File object if there's not as much data as the size says there is.